+++ src/internal.d.ts
@@ -1,5 +1,12 @@
-export interface InternalOptions {
+import type { PluginOptions } from '.'
+
+export interface InternalOptions extends PluginOptions {
   printer: Printer<any>
+
+  /**
+   * The package name to use when loading Tailwind CSS
+   */
+  tailwindPackageName?: string
 }
 
 export interface InternalPlugin {