+++ src/options.ts
@@ -1,6 +1,5 @@
 import type { RequiredOptions, SupportOption } from 'prettier'
 import type { Customizations } from './types'
-import './index'
 
 export const options: Record<string, SupportOption> = {
   tailwindConfig: {
@@ -54,6 +53,13 @@ export const options: Record<string, SupportOption> = {
     default: false,
     category: 'Tailwind CSS',
     description: 'Preserve duplicate classes inside a class list when sorting',
+  },
+
+  tailwindPackageName: {
+    type: 'string',
+    default: 'tailwindcss',
+    category: 'Tailwind CSS',
+    description: 'The package name to use when loading Tailwind CSS',
   },
 }