+++ tests/plugins.test.ts
@@ -1,5 +1,7 @@
 import { createRequire } from 'node:module'
+import dedent from 'dedent'
 import { test } from 'vitest'
+import { javascript } from './tests.js'
 import type { TestEntry } from './utils.js'
 import { format, no, pluginPath, t, yes } from './utils.js'
 
@@ -194,6 +196,19 @@ let tests: PluginTest[] = [
     },
   },
   {
+    plugins: ['@prettier/plugin-hermes'],
+    tests: {
+      hermes: javascript,
+    },
+  },
+  {
+    plugins: ['@prettier/plugin-oxc'],
+    tests: {
+      oxc: javascript,
+      'oxc-ts': javascript,
+    },
+  },
+  {
     plugins: ['@prettier/plugin-pug'],
     tests: {
       pug: [
@@ -221,23 +236,6 @@ let tests: PluginTest[] = [
         [
           `a.p-4.bg-blue-600(class='sm:p-0 md:p-4', href='//example.com') Example`,
           `a.bg-blue-600.p-4(class='sm:p-0 md:p-4', href='//example.com') Example`,
-        ],
-      ],
-    },
-  },
-  {
-    // NOTE: This plugin doesn't officially support Prettier v3 but it seems to work fine
-    plugins: ['prettier-plugin-import-sort'],
-    tests: {
-      babel: [
-        [
-          `
-            import './three'
-            import '@one/file'
-            import '@two/file'
-            export default function Foo() { return <div className="sm:p-0 p-4"></div> }
-          `,
-          `import '@one/file'\nimport '@two/file'\n\nimport './three'\n\nexport default function Foo() {\n  return <div className="p-4 sm:p-0"></div>\n}`,
         ],
       ],
     },
@@ -471,6 +469,77 @@ import Custom from '../components/Custom.astro'
         [
           `<div\n class={\`underline \n flex\`}></div>`,
           `<div\n  class={\`flex \n underline\`}\n></div>`,
+        ],
+
+        // Duplicates can be removed in simple attributes
+        [
+          `<div class="flex flex underline flex flex"></div>`,
+          `<div class="flex underline"></div>`,
+        ],
+
+        // Duplicates cannot be removed in string literals otherwise invalid
+        // code will be produced during printing.
+        [
+          `<div class={'flex underline flex'}></div>`,
+          `<div class={'flex flex underline'}></div>`,
+        ],
+
+        // Duplicates cannot be removed in template literals otherwise invalid
+        // code will be produced during printing.
+        [
+          `<div class={\`flex underline flex\`}></div>`,
+          `<div class={\`flex flex underline\`}></div>`,
+        ],
+      ],
+    },
+  },
+
+  // This test ensures that our plugin works with the multiline array, JSDoc,
+  // and import sorting plugins when used together.
+  //
+  // The plugins actually have to be *imported* in a specific order for
+  // them to function correctly *together*.
+  {
+    plugins: [
+      'prettier-plugin-multiline-arrays',
+      '@trivago/prettier-plugin-sort-imports',
+      'prettier-plugin-jsdoc',
+    ],
+    options: {
+      multilineArraysWrapThreshold: 0,
+      importOrder: ['^@one/(.*)$', '^@two/(.*)$', '^[./]'],
+      importOrderSortSpecifiers: true,
+    },
+    tests: {
+      babel: [
+        [
+          dedent`
+            import './three'
+            import '@two/file'
+            import '@one/file'
+
+            /**
+              * - Position
+              */
+            const position = {}
+            const arr = ['a', 'b', 'c', 'd', 'e', 'f']
+          `,
+          dedent`
+            import '@one/file'
+            import '@two/file'
+            import './three'
+
+            /** - Position */
+            const position = {}
+            const arr = [
+              'a',
+              'b',
+              'c',
+              'd',
+              'e',
+              'f',
+            ]
+          `,
         ],
       ],
     },